+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## [unreleased]
 
+ * Core: Fixed shutdown of fs watcher in development mode (hung sometimes)
+ * Metrics Collector: Reduced the number generated labels for *mashroom_http_request_* metrics.
+   This reduces the Prometheus load, because every label generates a new time series. Fixes #105.
  * Prometheus Exporter: Fixed a memory leak when the metrics were obtained via PM2 intercom
  * OpenID Connect Security Provider: Fixed the problem that authentication attempts could fail if the IDP redirected back
    very quickly, but the session was not persisted in the store yet